requirements.txt
prometheus-fastapi-instrumentator
pydantic==2.12.5
pytest==9.0.2
pytest-asyncio==1.3.0
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.28.1
requests>=2.31.0
aiohttp>=3.9.0
runwayml
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.19.0
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0
google-genai>=0.1.0
replicate>=0.20.0
scenepic
ffmpeg-python
opencv-python
av
# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.30.0
redis>=5.0.0
# Security & Authentication
firebase-admin>=6.0.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
# Observability - Phase 2
datadog>=0.45.0
prometheus_client>=0.19.0
google-cloud-logging>=3.8.0
jaeger-client==4.8.0
